replace passwd-user dependency
We need to remove the dependency on a GPL module, and passwd-user has a
nested dependency that is GPL licensed. Fortunately, it is only used
for functionality that we don't use, so it is trivial to simply
incorporate the parts of passwd user that we _do_ want directly and
simply add the MIT header.